Update C++ ver to behave exactly like the scripts

+++ c/fibcounter/test.cpp
@@ -1,23 +1,22 @@
 #include <stdio.h>
-
 int main(void) {
-    unsigned long long int a = 1;
-    unsigned long long int b = 1;
-    unsigned long long int x;
+    int a = 1;
+    int b = 1;
+    int x;
     x = a + b;
-    printf("%llu\n", a);
-    printf("%llu\n", b);
-    printf("%llu\n", x);
-    unsigned long long int max = 1000;
+    printf("%d\n", a);
+    printf("%d\n", b);
+    printf("%d\n", x);
+    int max = 1000;
     while (x < max) {
         a = b;
         b = x;
         x = a + b;
         if (x > max) {
-            printf("Done! Last number in sequence: %llu%s%llu\n", b, ", max number to look up to: ", max);
+            printf("Done! Last number in sequence: %d%s%d\n", b, ", max number to look up to: ", max);
         }
         else {
-            printf("%llu\n", x);
+            printf("%d\n", x);
         }
     }
 }
